+++ src/GeneticSampler.h
@@ -3,18 +3,18 @@ class GeneticSampler
 public:
 	GeneticSampler();
 
-	size_t 			get_nbRecombinations();
-    size_t 			get_T1_nbMuts();
-    size_t 			get_T2_nbMuts();
-    size_t 			get_T3_nbMuts();
-    size_t 			get_T4_nbMuts(const size_t nbGenerationsInBetween, const size_t left, const size_t right);
-    size_t 			get_T56_nbMuts();
-    size_t 			get_recombinationPosition();
-    size_t 			get_T1_mutationPosition();
-    size_t 			get_T2_mutationPosition();
-    size_t 			get_T3_mutationPosition();
-    size_t 			get_T4_mutationPosition(const size_t left, const size_t right);
-    size_t 			get_T56_mutationPosition();
+	uint32_t 			get_nbRecombinations();
+    uint32_t 			get_T1_nbMuts();
+    uint32_t 			get_T2_nbMuts();
+    uint32_t 			get_T3_nbMuts();
+    uint32_t 			get_T4_nbMuts(const uint32_t nbGenerationsInBetween, const uint32_t left, const uint32_t right);
+    uint32_t 			get_T56_nbMuts();
+    uint32_t 			get_recombinationPosition();
+    uint32_t 			get_T1_mutationPosition();
+    uint32_t 			get_T2_mutationPosition();
+    uint32_t 			get_T3_mutationPosition();
+    uint32_t 			get_T4_mutationPosition(const uint32_t left, const uint32_t right);
+    uint32_t 			get_T56_mutationPosition();
 
 
     void 			set_nbRecombinations(const double& rate);
@@ -35,11 +35,11 @@ private:
 
 	/* Set the poisson distributions first as I use the rates to assert that the cumulative sum of rate seem correct*/
 
-	std::poisson_distribution<size_t> poisson_nbRecombinations;
-	std::poisson_distribution<size_t> poisson_T1_nbMuts;
-	std::poisson_distribution<size_t> poisson_T2_nbMuts;
-	std::poisson_distribution<size_t> poisson_T3_nbMuts;
-	std::poisson_distribution<size_t> poisson_T56_nbMuts;
+	std::poisson_distribution<uint32_t> poisson_nbRecombinations;
+	std::poisson_distribution<uint32_t> poisson_T1_nbMuts;
+	std::poisson_distribution<uint32_t> poisson_T2_nbMuts;
+	std::poisson_distribution<uint32_t> poisson_T3_nbMuts;
+	std::poisson_distribution<uint32_t> poisson_T56_nbMuts;
 	double recombinationtotalRate;
 	double T1_mut_totalRate;
 	double T2_mut_totalRate;
